크루원 중 크루장 삭제 불가 로직 추가

See merge request s12-webmobile1-sub1/S12P11E204!196

+++ backend/butter/src/main/java/com/ssafy/butter/domain/crew/service/CrewServiceImpl.java
@@ -112,6 +112,9 @@ public class  CrewServiceImpl implements CrewService {
         validateCrewAdmin(crew, currentMember);
         Member member = memberService.findById(memberId);
         CrewMember crewMember = crewMemberRepository.findByCrewAndMember(crew, member).orElseThrow();
+        if (crewMember.getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Cannot delete crew admin");
+        }
         crewMemberRepository.delete(crewMember);
     }
